Extract FaviconFetcher into a separate file
